+++ tests/integrated/test-interchange-instability/2fluid.cxx
@@ -106,13 +106,13 @@ protected:
 
     BoutReal hthe0;
     if (mesh->get(hthe0, "hthe0") == 0) {
-      output.write("    ****NOTE: input from BOUT, Z length needs to be divided by %e\n",
+      output.write("    ****NOTE: input from BOUT, Z length needs to be divided by {:e}\n",
                    hthe0 / rho_s);
     }
 
     /************** NORMALISE QUANTITIES *****************/
 
-    output.write("\tNormalising to rho_s = %e\n", rho_s);
+    output.write("\tNormalising to rho_s = {:e}\n", rho_s);
 
     // Normalise profiles
     Ni0 /= Ni_x / 1.0e14;