+++ tests/unit/sys/test_expressionparser.cxx
@@ -36,7 +36,7 @@ public:
   clone(const std::list<std::shared_ptr<FieldGenerator>> args) override {
     if (args.size() != 2) {
       throw ParseException(
-          "Incorrect number of arguments to increment function. Expecting 2, got %zu",
+          "Incorrect number of arguments to increment function. Expecting 2, got {:d}",
           args.size());
     }
 
@@ -63,8 +63,8 @@ public:
   clone(const std::list<std::shared_ptr<FieldGenerator>> args) override {
     if (args.size() != 1) {
       throw ParseException(
-          "Incorrect number of arguments to increment function. Expecting 1, got %d",
-          static_cast<int>(args.size()));
+          "Incorrect number of arguments to increment function. Expecting 1, got {:d}",
+          args.size());
     }
 
     return std::make_shared<IncrementGenerator>(args.front());
@@ -90,8 +90,8 @@ public:
   clone(const std::list<std::shared_ptr<FieldGenerator>> args) override {
     if (args.size() != 0) {
       throw ParseException(
-          "Incorrect number of arguments to nullary function. Expecting 0, got %d",
-          static_cast<int>(args.size()));
+          "Incorrect number of arguments to nullary function. Expecting 0, got {:d}",
+          args.size());
     }
 
     return std::make_shared<NullaryGenerator>();
@@ -381,7 +381,7 @@ TEST_F(ExpressionParserTest, AddBinaryOp) {
 
 TEST(ParseExceptionTest, WhatTest) {
   try {
-    throw ParseException("%s", "test message");
+    throw ParseException("test message");
   } catch (ParseException &e) {
     std::string message{e.what()};
     EXPECT_NE(message.find("test message"), std::string::npos);