+++ CHANGELOG.md
@@ -14,6 +14,12 @@
 - Most factories and `create` methods standardised. Run
   `bin/bout-v5-factory-upgrader.py` on your physics models to update
   them [\#1842](https://github.com/boutproject/BOUT-dev/pull/1842)
+- We now use [fmt](https://fmt.dev) for all our string formatting,
+  instead of the printf-style formatting. This affects calls to
+  `Output`, `BoutException`/`ParseException`, `DataFile`,
+  `OptionsReader`, and `MsgStack`/`TRACE`. Run
+  `bin/bout-v5-format-upgrader.py` on your physics models to update
+  them. [\#1847](https://github.com/boutproject/BOUT-dev/pull/1847)
 
 ## [v4.3.0](https://github.com/boutproject/BOUT-dev/tree/v4.3.0) (2019-10-24)
 [Full Changelog](https://github.com/boutproject/BOUT-dev/compare/v4.2.3...v4.3.0)